% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feedback_loop.R
\name{CheckFeedbackLoop}
\alias{CheckFeedbackLoop}
\title{CheckFeedbackLoop}
\usage{
CheckFeedbackLoop(celltypes, wd_path, savepath, cts_of_interest)
}
\arguments{
\item{celltypes}{Vector, indicating the cell types in microenvironment.}

\item{savepath}{Character, the path where the result is stored.}

\item{cts_of_interest}{Vector or Character, the cell type your interest.}

\item{workdir}{Character, the path where the multilayer network is stored.}
}
\value{
list, containing the sender-receiver pair, and signals. The signals represent the multicellular feedback cruit between sender and receiver.
}
\description{
select multicellular feedback loop
}
