% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawAlluviumPlot}
\alias{DrawAlluviumPlot}
\title{DrawAlluviumPlot}
\usage{
DrawAlluviumPlot(
  InputDir,
  Cluster,
  ColorDB,
  Check = TRUE,
  top.n = 30,
  gtitle = "CCI",
  p_height = 9,
  p_width = 8
)
}
\arguments{
\item{InputDir}{Character, the path where the signals importance is stored.}

\item{Cluster}{Character, signals acitivity of upstream signal pairs of specific cluster are shown.}

\item{ColorDB}{Vector, containing the color of cell types, key siganls and others.}

\item{Check}{Logical, Whether to do check visualization.}

\item{top.n}{Numercial, Only relevant if Check=T. Number of top regulation between LR pairs and target genes to show according to the signals importance.}

\item{gtitle}{Character, the title of plot.}

\item{p_height}{Numercial, the height of plot.}

\item{p_width}{Numercial, the width of plot.}
}
\description{
Draw Alluvium Plot
}
