% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawBubblePlot}
\alias{DrawBubblePlot}
\title{DrawBubblePlot}
\usage{
DrawBubblePlot(respath, outputdir, ct, p_width, p_height)
}
\arguments{
\item{respath}{Character, the pathway where the all result is stored.}

\item{outputdir}{Character, the pathway to store the result .}

\item{p_width}{Numercial, the width of plot.}

\item{p_height}{Numercial, the height of plot.}
}
\description{
Draw Bubble Plot of LR signaling score
}
