% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawCirclePlot}
\alias{DrawCirclePlot}
\title{DrawCirclePlot}
\usage{
DrawCirclePlot(InputDir, Metric, ColorDB, gtitle = "CCI")
}
\arguments{
\item{InputDir}{Character, the path where the signals importance is stored.}

\item{Metric}{Character, indicating the metrics of edges in Network plot. Available options are: n_LRs, n_TGs, IM,  IM_norm,  mean_IM, mean_IM_norm}

\item{ColorDB}{Vector, indicating the colors of cell types.}

\item{gtitle}{Character, the title of plot.}
}
\description{
Draw Network Plot
}
