% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawCircosPlot}
\alias{DrawCircosPlot}
\title{DrawCircosPlot}
\usage{
DrawCircosPlot(InputDir, receiver, ColorDB)
}
\arguments{
\item{InputDir}{Character, the path where the signals importance is stored.}

\item{receiver}{Character, indicating the receiver cell type.}

\item{ColorDB}{Vector, indicating the colors of cell types.}
}
\description{
Draw Circos Plot
}
