% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawEdgeBundlingPlot}
\alias{DrawEdgeBundlingPlot}
\title{DrawEdgeBundlingPlot}
\usage{
DrawEdgeBundlingPlot(
  InputDir,
  Cluster,
  ColorDB,
  Check = TRUE,
  top.n = 50,
  gtitle = "CCI",
  p_height = 7.5,
  p_width = 7
)
}
\arguments{
\item{InputDir}{Character, the path where the signals acitivity is stored.}

\item{Cluster}{Character, signals acitivity of upstream signal pairs of specific cluster are shown.}

\item{ColorDB}{Vector, containing the color of cell types, key siganls and others.}

\item{Check}{Logical, Whether to do check visualization.}

\item{top.n}{Numercial, Only relevant if Check=T. Number of top regulation between ligands and receptor to show according to the signals acitivity.}

\item{gtitle}{Character, the title of plot.}

\item{p_height}{Numercial, the height of plot.}

\item{p_width}{Numercial, the width of plot.}
}
\description{
Draw EdgeBundling Plot
}
