% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawEnrichmentPlot}
\alias{DrawEnrichmentPlot}
\title{DrawEnrichmentPlot}
\usage{
DrawEnrichmentPlot(
  InputDir,
  Cluster,
  top.n = 3,
  gtitle = "Enrichment",
  p_height = 7.5,
  p_width = 7
)
}
\arguments{
\item{InputDir}{Character, the path where the multilayer network is stored.}

\item{Cluster}{Character, enrichment analysis of receptors's downstream target genes in specific cluster are shown.}

\item{top.n}{Numercial, Only relevant if Check=T. Number of top enrichment term to show according to the p value.}

\item{gtitle}{Character, the title of plot.}

\item{p_height}{Numercial, the height of plot.}

\item{p_width}{Numercial, the width of plot.}
}
\description{
Draw Enrichment Plot
}
