% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawHeatmapNetworkPlot}
\alias{DrawHeatmapNetworkPlot}
\title{DrawHeatmapNetworkPlot}
\usage{
DrawHeatmapNetworkPlot(
  InputDir,
  OutputDir,
  Metric,
  mycolor_ct,
  p_width,
  p_height
)
}
\arguments{
\item{InputDir}{Character, the path where the signals importance is stored.}

\item{Metric}{Character, indicating the metrics of edges in Network plot. Available options are: n_LRs, n_TGs, IM,  IM_norm,  mean_IM, mean_IM_norm}

\item{mycolor_ct}{Vector, indicating the colors of cell types.}

\item{p_width}{Numercial, the width of plot.}

\item{p_height}{Numercial, the height of plot.}

\item{outputdir}{Character, the pathway to save the result .}

\item{gtitle}{Character, the title of plot.}
}
\description{
Draw Network Plot
}
