% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawHeatmapPlot}
\alias{DrawHeatmapPlot}
\title{DrawHeatmapPlot}
\usage{
DrawHeatmapPlot(
  InputDir,
  Sender = NULL,
  Receiver = NULL,
  outputdir,
  p_height = 9,
  p_width = 8.4
)
}
\arguments{
\item{InputDir}{Character, the path where the stMLnet's result is stored.}

\item{Sender}{Character, the sender cell type, default = NULL.}

\item{Receiver}{Character, the receiver cell type, default = NULL.}

\item{outputdir}{Character, the path to save plot.}

\item{p_height}{Numercial, the height of plot.}

\item{p_width}{Numercial, the width of plot.}
}
\description{
Draw Heatmap Plot
}
