% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{DrawMLnetPlot}
\alias{DrawMLnetPlot}
\title{DrawMLnetPlot}
\usage{
DrawMLnetPlot(
  MLnetDir,
  ImportDir,
  Signal,
  ColorDB,
  Check = TRUE,
  top.n = 10,
  gtitle = "CCI",
  p_height = 4.5,
  p_width = 7
)
}
\arguments{
\item{MLnetDir}{Character, the path where the multilayer network is stored.}

\item{ImportDir}{Character, the path where the signals importance is stored.}

\item{Signal}{Character, downstream multilayer network of specific signal are shown.}

\item{ColorDB}{Vector, indicating the colors of key siganls including Ligands, Recetors, TFs and Target genes.}

\item{Check}{Logical, Whether to do check visualization.}

\item{top.n}{Numercial, Only relevant if Check=T. Number of top regulation between LR pairs and target genes to show according to the signals importance.}

\item{gtitle}{Character, the title of plot.}

\item{p_height}{Numercial, the height of plot.}

\item{p_width}{Numercial, the width of plot.}
}
\description{
Draw MLnet Plot
}
