% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_network.R
\name{getBarList}
\alias{getBarList}
\title{getBarList}
\usage{
getBarList(Aclu, BarCluTable)
}
\arguments{
\item{Aclu}{cluster of interset}

\item{BarCluTable}{Matrix, The annotation matrix for cell type annotation. The first column is barcode and the second is cell type.}
}
\value{
List. The first sublist is the barcodes of cluster of interset, the second sublist is the barcodes of other cluster.
}
\description{
get barcode list of cluster of interset and other cluster.
}
