% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_network.R
\name{getCellPairMLnet}
\alias{getCellPairMLnet}
\title{getCellPairMLnet}
\usage{
getCellPairMLnet(inputs, ligclu, recclu, databases)

getCellPairMLnet(inputs, ligclu, recclu, databases)
}
\arguments{
\item{inputs}{List, the parameters and data inputted by current project.}

\item{ligclu}{Character, cell types that sending signals.}

\item{recclu}{Character, cell types that receiving signals.}

\item{databases}{List, The prior database used by running jobs for now. Databases includes Ligand-Receptor interactions (LigRec.DB), Receptor-TF interactions (RecTF.DB) and TF-Target interactions (TFTG.DB).}
}
\value{
List, The first sublist (mlnet) is the multilayer signal network between ligclu and recclu, the second sublist (detail) is the detail of cell communication between ligclu and recclu.

List, The first sublist (mlnet) is the multilayer signal network between ligclu and recclu, the second sublist (detail) is the detail of cell communication between ligclu and recclu.
}
\description{
obtain the multilayer signal network of specific cell pair (LigClu-RecClu).

obtain the multilayer signal network of specific cell pair (LigClu-RecClu).
}
