% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signal_activity.R
\name{getCPSiganlActivity}
\alias{getCPSiganlActivity}
\title{getCPSiganlActivity}
\usage{
getCPSiganlActivity(
  exprMat,
  distMat,
  annoMat,
  LRpairs,
  TGs,
  receiver,
  sender = NULL
)
}
\arguments{
\item{exprMat}{Matrix, The expression matrix with rows as genes (gene symbols) and columns as cells.}

\item{distMat}{Matrix, The distance matrix represents the distance between two cells (e.g., Euclidean distance).}

\item{annoMat}{Matrix, The annotation matrix for cell type annotation. The first column is barcode and the second is cell type.}

\item{LRpairs}{List, upstream signal pairs (ligand/receptor) in the multilay signal network between sender and receiver.}

\item{TGs}{Vector, downstream target genes in the multilay signal network between sender and receiver.}

\item{receiver}{Character, cell type that sending signals.}

\item{sender}{Character, cell type that receiving signals.}
}
\value{
List. The sublist contains the activity of upstream signal pairs (LRs_score) and expression of downstream target genes (TGs_expr) between sender and receiver.
}
\description{
calculate the upstream signals activity in the multilay signal network of the specific celltype pair
}
