% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_network.R
\name{getDiffExpGene}
\alias{getDiffExpGene}
\title{getDiffExpGene}
\usage{
getDiffExpGene(inputs)
}
\arguments{
\item{inputs}{List, the parameters and data inputted by current project.}
}
\value{
List, inputs object contains the differently expressed genes.
}
\description{
obtain the differently expressed genes using FindMarkers function in Seurat package.
}
