% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_network.R
\name{getHighExpGene}
\alias{getHighExpGene}
\title{getHighExpGene}
\usage{
getHighExpGene(inputs)
}
\arguments{
\item{inputs}{List, the parameters and data inputted by current project.}
}
\value{
List, inputs object contains the highy expressed genes.
}
\description{
obtain the highly expressed genes according the detection and mean expression.
}
