% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signal_importance.R
\name{getOOBPred}
\alias{getOOBPred}
\title{getOOBPred}
\usage{
getOOBPred(rfmodel, data_X)
}
\arguments{
\item{rfmodel}{an ranger object, the trained model of specific target genes.}

\item{data_X}{Matrix, the activity of upstream signal pairs.}
}
\value{
Matrix, the predicted activity of upstream signal pairs.
}
\description{
get predictions for each tree in trained model based on OOB data
}
