% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signal_importance.R
\name{getPartImport}
\alias{getPartImport}
\title{getPartImport}
\usage{
getPartImport(rfmodel, data_X, data_y, LRTab)
}
\arguments{
\item{rfmodel}{an ranger object, the trained model of specific target genes.}

\item{data_X}{Matrix, the activity of upstream signal pairs.}

\item{data_y}{Vector, the expression of downstream target genes.}

\item{LRTab}{Matrix, the upstream signal pairs.}
}
\value{
Matrix, the importance of partial variable (signals) corresponding the specific target genes
}
\description{
get importance of partial variable (signals) by shuffle in OOB data
}
