% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feedback_loop.R
\name{plot_feedback_loop}
\alias{plot_feedback_loop}
\title{plot_feedback_loop}
\usage{
plot_feedback_loop(floop, save_path, vertex.size, p_width, p_height)
}
\arguments{
\item{floop}{table, containing the sender, receiver, and signals. The signals represent the multicellular feedback cruit between sender and receiver.}

\item{vertex.size}{Numercial, the size of nodes.}

\item{p_width}{Numercial, the width of plot.}

\item{p_height}{Numercial, the height of plot.}

\item{savepath}{Character, the path where the result is stored.}
}
\description{
plot multicellular feedback loop
}
