% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feedback_loop.R
\name{plot_feedback_network}
\alias{plot_feedback_network}
\title{plot_feedback_network}
\usage{
plot_feedback_network(
  wd_path,
  ct1,
  ct2,
  ct1_key,
  ct2_key,
  vertex.size,
  rescale = FALSE,
  do.check = TRUE,
  p_width = 10,
  p_height = 6
)
}
\arguments{
\item{wd_path}{Character, the path where the stMLnet inference result is stored..}

\item{ct1}{Character, sender cell.}

\item{ct2}{Character, receiver cell.}

\item{ct1_key}{Character, the signaling key gene of sender cell.}

\item{ct2_key}{Character, the signaling key gene of receiver cell.}

\item{vertex.size}{Numercial, the size of nodes.}

\item{rescale}{Character, Whether to automatically scale.}

\item{do.check}{Character, check the result for better visualization.}

\item{p_width}{Numercial, the width of plot.}

\item{p_height}{Numercial, the height of plot.}
}
\description{
plot multilayer signaling feedback network
}
