% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cell_communication.R
\name{PrepareColorDB}
\alias{PrepareColorDB}
\title{PrepareColorDB}
\usage{
PrepareColorDB(CellTypes, ColorDB = NULL)
}
\arguments{
\item{CellTypes}{Vector, indicating the cell types in microenvironment.}

\item{ColorDB}{Vector, indicating the colors of cell types.}
}
\value{
List, containing the color of cell types, key siganls and others.
}
\description{
Prepare color database for visualization
}
