% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signal_activity.R
\name{runImputation}
\alias{runImputation}
\title{getCPSiganlActivity}
\usage{
runImputation(exprMat, use.seed = TRUE, seed = 4321)
}
\arguments{
\item{exprMat}{Matrix, The expression matrix with rows as genes (gene symbols) and columns as cells.}

\item{use.seed}{Logical, Whether to set a seed.}

\item{seed}{Only relevant if use.seed=T, set the seed for the random process.}
}
\value{
Matrix, the imputed expression matrix.
}
\description{
get The imputation of expression matrix using the ALRA method in Seurat
}
