% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_network.R
\name{runMLnet}
\alias{runMLnet}
\title{runMLnet}
\usage{
runMLnet(
  ExprMat,
  AnnoMat,
  LigClus = NULL,
  RecClus = NULL,
  Normalize = T,
  NormMethod = "LogNormalze",
  OutputDir = NULL,
  Databases = NULL,
  RecTF.method = "Fisher",
  TFTG.method = "Fisher",
  logfc.ct = 0.1,
  pct.ct = 0.05,
  pval.ct = 0.05,
  expr.ct = 0.1,
  TGList = NULL,
  LigList = NULL,
  RecList = NULL
)
}
\arguments{
\item{ExprMat}{Matrix, The expression matrix with rows as genes (gene symbols) and columns as cells.}

\item{AnnoMat}{Matrix, The annotation matrix for cell type annotation. The first column is barcode and the second is cell type.}

\item{LigClus}{Vector, cell types that sending signals.}

\item{RecClus}{Vector, cell types that receiving signals.}

\item{Normalize}{Logical, Whether to do normalization.}

\item{NormMethod}{Character, Denotes which normalization method to use. Available options are: LogNormalze, SCTransform. See Seurat package for more details.}

\item{OutputDir}{Character, The output path of the currently running job where temporary and final results will be saved.}

\item{Databases}{List, The prior database used by running jobs for now. Databases includes Ligand-Receptor interactions (LigRec.DB), Receptor-TF interactions (RecTF.DB) and TF-Target interactions (TFTG.DB).}

\item{RecTF.method}{Character, Denotes the strategy for filtering downstream pairing signals (Receptor-TF).  Available options are: Fisher(default, meaning Fisher exact test) and Search (meaning searching in database).}

\item{TFTG.method}{Character, Denotes the strategy for filtering downstream pairing signals (TF-Target).  Available options are: Fisher(default, meaning Fisher exact test) and Search (meaning searching in database).}

\item{logfc.ct}{Numercial, Screening threshold for FindMarkers in Seurat. The default setting is 0.1.}

\item{pct.ct}{Numercial, Screening threshold for FindMarkers in Seurat. The default setting is 0.05.}

\item{pval.ct}{Numercial, Screening threshold for FindMarkers in Seurat. The default setting is 0.05.}

\item{expr.ct}{Numercial, Screening threshold for high expressed gene in groups of cells. Default is 0.05.}

\item{TGList}{List, The target genes of interest in groups of cells (RecClus).}

\item{LigList}{List, The potential ligands in groups of cells (LigClus).}

\item{RecList}{List, The potential receptors in groups of cells (RecClus).}
}
\value{
List. The first sublist (mlnets) is the multilayer signal network between LigClus and RecClu, the second sublist (detail) is the detail of cell communication between LigClus and RecClu.
}
\description{
Creates an object that stores the multilayer signal network between LigClus and RecClus
}
