% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multilayer_network.R
\name{runNormalize}
\alias{runNormalize}
\title{runNormalize}
\usage{
runNormalize(inputs, method = c("LogNormalze", "SCTransform"))
}
\arguments{
\item{inputs}{List, the parameters and data inputted by current project.}

\item{method}{Character, Normalization method, available options are: LogNormalze, SCTransform. See Seurat package for more details.}
}
\value{
List, inputs object contains the normalized data.
}
\description{
Normalize the expression data using function in Seurat package.
}
