% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signal_importance.R
\name{runRFModel}
\alias{runRFModel}
\title{runRFModel}
\usage{
runRFModel(
  trainx,
  trainy,
  auto_para = TRUE,
  n.trees = 500,
  n.trys = 10,
  tree.method = "variance",
  node.size = 5,
  nPert = 10
)
}
\arguments{
\item{trainx}{Matrix, the activity of upstream signal pairs.}

\item{trainy}{Vector, the expression of downstream target genes.}

\item{auto_para}{Logical, Whether to do automatic optimization parameter.}

\item{n.trees}{Numercial, number of trees in random forests model, see Seurat package for more details.}

\item{n.trys}{Numercial, Only relevant if AutoPara=F. Number of variables to possibly split at in each node, see Seurat package for more details.}

\item{tree.method}{Character, Only relevant if AutoPara=F. Splitting rule, see Seurat package for more details.}

\item{node.size}{Numercial, Only relevant if AutoPara=F. Minimal node size, see Seurat package for more details.}

\item{nPert}{Numercial, Number of shuffle.}
}
\value{
List, The first sublist (model) is the trained model of specific target genes. The second sublist (df_IM) is the importance of upstream signal pairs corresponding the specific target genes and the third sublist (df_pIM) is the importance of upstream signals corresponding the specific target genes.
}
\description{
train a random forest model for specific target genes to calculate the importance of upstream signal pairs or signals
}
