% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feedback_loop.R
\name{SaveFeedbackLoop}
\alias{SaveFeedbackLoop}
\title{SaveFeedbackLoop}
\usage{
SaveFeedbackLoop(key_of_inter, cts_of_interest, savepath)
}
\arguments{
\item{key_of_inter}{data.frame, the result of 'CheckFeedbackLoop' function.}

\item{cts_of_interest}{Vector or Character, the cell type your interest.}

\item{savepath}{Character, the path where the result is stored.}
}
\value{
Table, containing the sender, receiver, and signals. The signals represent the multicellular feedback cruit between sender and receiver.
}
\description{
save multicellular feedback loop as table
}
