% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signal_importance.R
\name{shufflePartVars}
\alias{shufflePartVars}
\title{shufflePartVars}
\usage{
shufflePartVars(data_X, i, LRTab)
}
\arguments{
\item{data_X}{Matrix, the activity of upstream signal pairs.}

\item{i}{i-th partial variables to be shuffled.}

\item{LRTab}{Matrix, the upstream signal pairs.}
}
\value{
Matrix, the activity of upstream signal pairs after shuffling
}
\description{
shuffle the activity of upstream signal pairs that involve i-th partial variables
}
