% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feedback_loop.R
\name{CheckFeedbackLoop}
\alias{CheckFeedbackLoop}
\title{check_feedback_loop}
\usage{
CheckFeedbackLoop(celltypes, workdir, savepath, cts_of_interest)

CheckFeedbackLoop(celltypes, workdir, savepath, cts_of_interest)

CheckFeedbackLoop(celltypes, workdir, savepath, cts_of_interest)

CheckFeedbackLoop(celltypes, workdir, savepath, cts_of_interest)
}
\arguments{
\item{celltypes}{Vector, indicating the cell types in microenvironment.}

\item{workdir}{Character, the path where the multilayer network is stored.}

\item{savepath}{Character, the path where the result is stored.}

\item{cts_of_interest}{Vector or Character, the cell type your interest.}
}
\value{
Table, containing the sender, receiver, and signals. The signals represent the multicellular feedback cruit between sender and receiver.

Table, containing the sender, receiver, and signals. The signals represent the multicellular feedback cruit between sender and receiver.

Table, containing the sender, receiver, and signals. The signals represent the multicellular feedback cruit between sender and receiver.

Table, containing the sender, receiver, and signals. The signals represent the multicellular feedback cruit between sender and receiver.
}
\description{
select multicellular feedback loop

select multicellular feedback loop

select multicellular feedback loop

select multicellular feedback loop
}
