% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_signal_activity.R
\name{getSiganlActivity}
\alias{getSiganlActivity}
\title{getSiganlActivity}
\usage{
getSiganlActivity(
  ExprMat,
  DistMat,
  AnnoMat,
  MulNetList,
  Receiver,
  Sender = NULL,
  OutputDir = NULL
)
}
\arguments{
\item{ExprMat}{Matrix, The expression matrix with rows as genes (gene symbols) and columns as cells.}

\item{DistMat}{Matrix, The distance matrix represents the distance between two cells (e.g., Euclidean distance).}

\item{AnnoMat}{Matrix, The annotation matrix for cell type annotation. The first column is barcode and the second is cell type.}

\item{MulNetList}{List, The sublist is the multilayer signal network between of specific celltype pair.}

\item{Receiver}{Character, cell type that sending signals.}

\item{Sender}{Vector, cell types that receiving signals.}

\item{OutputDir}{Character, The output path of the currently running job where temporary and final results will be saved.}
}
\value{
List, The sublist contains the activity of upstream signal pairs and expression of downstream target genes in specific celltype pair.
}
\description{
calculate the upstream signals activity in the multilay signal network of cell communication
}
